feat: add nodes
Nodes are a data structure that dendriform uses to keep track of unique items within the data shape,
despite those pieces of data potentially moving around over time. Each node has a unique id, and each
node knows the ids of its children and the property name or index required to access them.

Changes in nodes are often produced in sympathy with changes in the value shape. These make heavy use
of immer patches (and dendriform-immer-patch-optimiser) to track movements and maintain referential integrity.